feat: 일정과 댓글의 연관관계 설정

+++ src/main/java/com/sparta/bizee/Repository/ScheduleRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface ScheduleRepository extends JpaRepository<Schedule, Integer> {
+public interface ScheduleRepository extends JpaRepository<Schedule, Long> {
     // 쿼리 메소드
     // 작성일자 내림차순
     List<Schedule> findAllByOrderByCreationDateDesc();
